+++ recipes/install-elasticsearch.rb
@@ -39,6 +39,7 @@ include_recipe "mh-opsworks-recipes::update-package-repo"
 pin_package("elasticsearch", "#{es_major_version}.*")
 pin_package("python-elasticsearch-curator", "#{curator_major_version}.*")
 install_package("openjdk-7-jdk openjdk-7-jre elasticsearch python-elasticsearch-curator")
+configure_cloudwatch_log("elasticsearch", "/var/log/elasticsearch/*.log", "%Y-%m-%d %H:%M:%S")
 
 service 'elasticsearch' do
   supports :restart => true